backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
langchain==0.1.0
langchain-openai==0.0.2
tiktoken==0.5.2
orjson==3.9.10
pydantic==2.5.0
python-multipart==0.0.6
cors==1.0.1